refactor moving prisma module to common and Date scalar
