+++ src/auth/entities/auth.entity.ts
@@ -1,6 +1,12 @@
-import { Field, HideField, ID, ObjectType } from '@nestjs/graphql';
+import {
+  Field,
+  GraphQLISODateTime,
+  HideField,
+  ID,
+  ObjectType,
+} from '@nestjs/graphql';
 import { Auth } from '@prisma/client';
-import { UserEntity } from '../../users/entities/user.entity'; // Import the UserEntity we created earlier
+import { UserEntity } from '../../users/entities/user.entity';
 
 @ObjectType()
 export class AuthEntity implements Auth {
@@ -16,10 +22,10 @@ export class AuthEntity implements Auth {
   @HideField()
   readonly refreshToken: string;
 
-  @Field(() => Date)
+  @Field(() => GraphQLISODateTime)
   readonly refreshExpiresAt: Date;
 
-  @Field(() => Date)
+  @Field(() => GraphQLISODateTime)
   readonly createdAt: Date;
 
   @Field(() => UserEntity)
