+++ src/brands/brands.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { BrandsService } from './services/brands.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../common/prisma/prisma.service';
 
 @Module({
   providers: [BrandsService, PrismaService],
