+++ src/categories/categories.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CategoriesService } from './services/categories.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../common/prisma/prisma.service';
 
 @Module({
   providers: [CategoriesService, PrismaService],
