+++ src/categories/entitites/category.entity.ts
@@ -1,4 +1,4 @@
-import { Field, ID, ObjectType } from '@nestjs/graphql';
+import { Field, GraphQLISODateTime, ID, ObjectType } from '@nestjs/graphql';
 import { Category } from '@prisma/client';
 import { ProductEntity } from '../../products/entities/product.entity';
 
@@ -10,10 +10,10 @@ export class CategoryEntity implements Category {
   @Field(() => String)
   readonly name: string;
 
-  @Field(() => Date)
+  @Field(() =>GraphQLISODateTime)
   readonly createdAt: Date;
 
-  @Field(() => Date)
+  @Field(() =>GraphQLISODateTime)
   readonly updatedAt: Date;
 
   @Field(() => [ProductEntity], { nullable: true })
