+++ src/common/guards/admin.guard.ts
@@ -5,7 +5,7 @@ import {
   Injectable,
   UnauthorizedException,
 } from '@nestjs/common';
-import { PrismaService } from '../../prisma/prisma.service';
+import { PrismaService } from '../prisma/prisma.service';
 import { GqlExecutionContext } from '@nestjs/graphql';
 
 @Injectable()
