+++ src/dataloader/dataloader.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CategoriesService } from '../categories/services/categories.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../common/prisma/prisma.service';
 import { DataloaderService } from './services/dataloader.service';
 import { BrandsService } from '../brands/services/brands.service';
 import { VariantsService } from '../variants/services/variants.service';
